Balance and Pending Transactions

+++ lib/features/home/domain/entity/balance_entity.dart
@@ -2,7 +2,7 @@ import 'package:equatable/equatable.dart';
 
 class BalanceEntity extends Equatable {
   final int id;
-  final double walletBalance;
+  final int walletBalance;
   final int walletPin;
 
   const BalanceEntity({
